Connect new node
